refactor(windows-lib): move capture/texture code to shared module
