+++ crates/windows-lib/src/shared/mod.rs
@@ -1,4 +1,6 @@
 mod workspace_clients;
 mod workspaces;
+mod capture_utils;
 
 pub use workspaces::*;
+pub use capture_utils::*;
