+++ crates/windows-lib/src/switch/root.rs
@@ -1,10 +1,10 @@
 use crate::data::{SortConfig, collect_data};
 use crate::next::{find_next_client, find_next_workspace};
-use crate::shared::{Workspaces, WorkspacesInit, WorkspacesInput};
+use crate::shared::{Workspaces, WorkspacesInit, WorkspacesInput, refresh_captures};
 use crate::switch::clients::{Clients, ClientsInit, ClientsInput};
-use core_lib::{Active, ByFirst, ClientId, Direction, HyprlandData, SWITCH_NAMESPACE};
+use core_lib::{Active, ByFirst, Direction, HyprlandData, SWITCH_NAMESPACE};
 use exec_lib::switch::{switch_client, switch_workspace};
-use exec_lib::wayland_capture::{CaptureManager, CaptureMode, CaptureOutput, ObjectId};
+use exec_lib::wayland_capture::{CaptureManager, CaptureMode};
 use gtk4_layer_shell::{KeyboardMode, Layer, LayerShell};
 use regex::Regex;
 use relm4::adw::glib::ControlFlow;
@@ -14,7 +14,6 @@ use relm4::adw::prelude::*;
 use relm4::gtk::gdk::Key;
 use relm4::gtk::{EventControllerKey, Orientation, SelectionMode};
 use relm4::prelude::*;
-use std::os::fd::AsRawFd;
 use std::collections::HashMap;
 use std::time::Duration;
 use tracing::{debug, error, trace};
@@ -200,63 +199,12 @@ impl SimpleComponent for SwitchRoot {
             }
             SwitchRootInput::RefreshThumbnails => {
                 let Some(mgr) = &mut self.capture_manager else { return; };
-                let _ = mgr.dispatch_pending();
-
-                mgr.drain_closed();
-
-                let _ = mgr.drain_new();
-
                 let display = RootExt::display(&self.window);
-
-                let capture_map: HashMap<ClientId, ObjectId> = mgr.capture_ids()
-                                                                  .into_iter()
-                                                                  .filter_map(|oid| mgr.client_id(&oid).map(|cid| (cid, oid)))
-                                                                  .collect();
+                let mut captures = refresh_captures(mgr, &display);
                 for (idx, item) in self.clients_only.iter().enumerate() {
-                    let Some(obj_id) = capture_map.get(&item.id) else { continue; };
-
-                    if mgr.is_failed(&obj_id) {
-                        // TODO: gérer l'erreur
-                        let _ = mgr.capture_next(&obj_id);
+                    if let Some(texture) = captures.remove(&item.id) {
+                        self.clients_only.send(idx, ClientsInput::UpdateThumbnail(texture));
                     }
-
-                    if !mgr.is_ready(&obj_id) { continue; }
-
-                    let texture = match mgr.take_output(&obj_id) {
-                        Ok(CaptureOutput::Dmabuf(dmabuf)) => {
-                            match unsafe {
-                                gtk::gdk::DmabufTextureBuilder::new().set_display(&display)
-                                                                     .set_width(dmabuf.width)
-                                                                     .set_height(dmabuf.height)
-                                                                     .set_fourcc(dmabuf.fourcc)
-                                                                     .set_modifier(dmabuf.modifier)
-                                                                     .set_n_planes(1)
-                                                                     .set_fd(0, dmabuf.fd.as_raw_fd())
-                                                                     .set_stride(0, dmabuf.stride)
-                                                                     .set_offset(0, 0)
-                                                                     .build()
-                            } {
-                                // TODO: gérer l'erreur
-                                Ok(t) => t.upcast(),
-                                Err(_) => {
-                                    continue;
-                                }
-                            }
-                        }
-                        Ok(CaptureOutput::Shm(shm_result)) => {
-                            let bytes = gtk::glib::Bytes::from(&shm_result.pixels);
-                            gtk::gdk::MemoryTexture::new( shm_result.width as i32
-                                                        , shm_result.height as i32
-                                                        , gtk::gdk::MemoryFormat::B8g8r8a8Premultiplied
-                                                        , &bytes
-                                                        , shm_result.stride as usize
-                                                        ).upcast()
-                        }
-                        Err(_) => { continue; }
-                    };
-                    self.clients_only.send(idx, ClientsInput::UpdateThumbnail(texture));
-
-                    let _ = mgr.capture_next(&obj_id);
                 }
             }
         }
